images/tools/database-copy3/pyproject.toml
[tool.poetry]
name = "database-copy3"
version = "0.1.0"
description = ""
authors = ["Kate Sugden <kate.sugden@madetech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
boto3 = "^1.35.48"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
